Groundwork for filter history display.
That is, it doesn't work yet, but I'm working on it and I need to sleep. I've rigged it up to make up to five different TextViews that display filters you've applied. It crashes after five, but that's just because I haven't set up an interface to tell the user not to do more than five.

Oh, and the TextViews are all in the top left corner. I'll be fixing that tomorrow.